BlogProject/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-tinymce==3.4.0
Pillow==9.1.1
sqlparse==0.4.2
tzdata==2022.1
